requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.1.0
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==4.2.5
django-admin==2.0.2
django-cors-headers==4.0.0
django-environ==0.9.0
django-excel-response2==3.0.6
django-six==1.0.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
excel-base==1.0.4
gunicorn==20.1.0
idna==3.4
isoweek==1.3.3
oauthlib==3.2.2
psycopg2==2.9.6
psycopg2-binary==2.9.8
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
screen==1.0.1
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
TimeConvert==3.0.13
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.15
xlwt==1.3.0